+++ Assets/CommonCore/Core/CoreParams.cs
@@ -33,8 +33,10 @@ namespace CommonCore
         private static WindowsPersistentDataPath PersistentDataPathWindows = WindowsPersistentDataPath.Roaming;
         private static bool CorrectWindowsLocalDataPath = false; //if set, use AppData/Local/* instead of AppData/LocalLow/* for LocalDataPath
         private static bool UseGlobalScreenshotFolder = true; //ignored on UWP and probably other platforms
-        public static bool UseMigrationBackups { get; private set; } = true; //if set, will save backups before migrating
-
+		
+        public static bool UseSystemMigrationBackups { get; private set; } = true; //if set, will save backups before migrating, affects PersistState and ConfigState
+        public static bool UseSaveMigrationBackups { get; private set; } = true; //affects game saves
+		
         public static bool SetSafeResolutionOnExit { get; private set; } = true;
         public static Vector2Int SafeResolution { get; private set; } = new Vector2Int(1280, 720);
         public static bool SetNativeResolutionOnFirstRun { get; private set; } = true;