+++ CHANGELOG.md
@@ -527,3 +527,4 @@
 * CharacterModel EquippedDictionaryProxy no longer breaks when trying to retrieve the equipped item in a slot that has no item equipped
 * Quicksaves are now checked for game version and will not load if version mismatched (controlled by EnforceQuicksaveVersionMatching)
 * Warning is now displayed when attempting to manually load a save from a previous game version
+* UseMigrationBackups is now split into UseSaveMigrationBackups (affects game saves) and UseSystemMigrationBackups (affects ConfigState and PersistState)